+++ src/pages/api/users/[uid].js
@@ -7,6 +7,7 @@ import {
 	where,
 	deleteDoc,
 	getDocs,
+	getDoc,
 	deleteDoc,
 } from 'firebase/firestore';
 import APIRouteHelper from '@/common/api/APIRouteHelper';
@@ -32,7 +33,12 @@ async function onGET(req, res) {
 	try {
 		const usersRef = collection(db, 'users');
 		const usersQuery = query(usersRef, where('id', '==', uid));
-		const usersDocs = await getDocs(usersQuery);
+		const usersDocs = await getDoc(usersQuery);
+		if (usersDocs.exists() === false) {
+			res.status(404).json({ error: 'User not found' });
+			return;
+		}
+
 		const data = usersDocs.docs[0].data();
 
 		res.status(200).json({ data });
@@ -45,6 +51,12 @@ async function onDELETE(req, res) {
 	try {
 		const usersRef = collection(db, 'users');
 		const usersQuery = query(usersRef, where('id', '==', uid));
+		const usersDocs = await getDoc(usersQuery);
+
+		if (usersDocs.docs.length === 0) {
+			res.status(404).json({ error: 'User not found' });
+			return;
+		}
 		await deleteDoc(doc(usersQuery));
 		//  await deleteDoc(usersRef, where('id', '==', id));
 		res.status(200).json('User deleted');
@@ -55,17 +67,19 @@ async function onDELETE(req, res) {
 
 async function onPUT(req, res) {
 	try {
-		let randonLocalidad = Math.floor(Math.random() * localidades.length);
-		let localidadId = localidades[randonLocalidad].id;
-		const docRef = await addDoc(collectionRef, {
-			obj,
-			localidad: localidadId,
-			createdAt: new Date(),
-		});
-		console.log(docRef);
-		const id = docRef.id;
-		await updateDoc(doc(collectionRef, id), { id });
-		console.log(`Document created with ID: ${id}`);
+		let id = uid;
+		const usersRef = collection(db, 'users');
+		const usersDocs = await getDoc(query(usersRef, where('id', '==', id)));
+		if (usersDocs.exists() === false) {
+			res.status(404).json({ error: 'User not found' });
+			return;
+		}
+
+		await updateDoc(doc(usersRef, id), req.body);
+		console.log(
+			`User updated created with ID: ${id} and with body: ${req.body}`
+		);
+		res.status(200).json({ data: 'User updated' });
 	} catch (error) {
 		res.status(404).json({ error });
 	}
